NRF_RADIO: Handle more gracefully header errors
If a header error occurs while receiving a packet, the RADIO
model could set a timer in the past which caused a warning,
this was due to a bad calculation of the packet timing.
This commit fixes this, by always assuming the received packet
lenght is the correct one (even during header errors), which
always results in the packet being received until the "correct"
CRC end (like if there was just a CRC error)
This is a good enough approximation for the purpose of this model,
specially given that we do not really know how the real Nordic
HW behaves depending on where the bit errors in the header are.

Signed-off-by: Alberto Escolar Piedras <alpi@oticon.com>