HAL: AAR: Add missing copyright and license
This file was contributed in
d6ecc005a715119107a26f41ce899fb9892f8fc5

But the license and copyright header was missing.

The contribution was properly signed off, and therefore
properly licensed.
But for clarity and consistency let's add the header
to the file.

The code itself is mostly a copy of content with from
another file with Copyright (c) 2017 Oticon A/S with minor
changes from Nordic, therefore add both copyrights to it

Signed-off-by: Alberto Escolar Piedras <alpi@oticon.com>