Model a nrf52833 instead of a nrf52832
Motivation for the change:

The nrf52833 is for the purpose of these models practically the
same as a 52832, but it includes the HEADER_MASK register in the
AES_CCM module which is needed for proper encrypted BIS support.

Changing the version of the HW we model can cause minor trouble
for users, i.e. require them to change the configuration of their SW,
but at this point is believed most users will desire sooner or
later full BLE ISO support.

Change:

* Changed version we compile as from NRF52832_XXAA to NRF52833_XXAA
* Number of interrupts increased to 47 (32 was too low even for 52832)
* Updated all documentation references to point to 52833 docs
* Updated description of missing features (no new 52833 features
  implemented so far)

******************************************************************
Expect backwards/forward compatibility changes from/to this commit
******************************************************************

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>