AES_CCM: Utilize newer ext_libCryptv1 v1.2 API
The old API (used by this commit) which applied the old
AAD 0xE3 mask on top of the HEADERMASK.
This is ok for BLE (so far) as all masks
are at least more restrictive than that. But it is
in general not correct.

So let's update to the new API

Note: This commit requires ext_libCryptov1 v1.2
or newer.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>