AES_CCM: Add HEADERMASK functionality
This change is required for proper BT LE ISO support
(BIS encryption with control subevents present).

Note: The change has been done relaying on the old
ext_libCryptov1 v1.1 API, even though the new APIs
introduced in v1.2 would have been better.

The old API (used by this commit) will apply the old
AAD 0xE3 mask on top of the HEADERMASK.
This is ok for BLE (so far) as all masks
are at least more restrictive than that. But it is
in general not correct.

Creating a dependency for libCrypto v1.2 is considered
too much of a hassle for users, as even Zephyr docker
image is not updated to use it yet.

The proper change is parked in a separate commit/branch:
HEADERMASK_PROPER
In a few months we should consider taking it in.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>