RADIO: Improve T_IFS support, and fix for low latency mode
a) Fix for the HW automatically generated TIFS:
One condition which disables this HW feature was missing,
which lead to it being enabled when it shouldnt in some cases.
=> Fix it

b) Fix the DISABLED=>[TR]XEN shortcuts when automatic HW IFS is not used.
The shortcuts were missing in this case.

c) Postpone Tx request to the Phy until the last moment:
That is, do the request Tx_chain_delay later.
As otherwise, if another (pending) HW event was scheduled at the exact
same time as the Tx start it would cause the request to have an abort
reevaluation before the Tx start, which is a violation of the
protocol to the phy.
This issue was triggered, in low latency mode, where a TIMER event
was used to trigger the Tx start task, but could have been triggered
by other concurrent HW events.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>