RADIO: Bugfix if DISABLE or STOP during first us of TxStart
If a DISABLE or STOP was received during the first TxChainDelay (1us)
after a Tx START, the task would have been missed.
Fix it.

The issue was introduced in cee41373eb0bb720a370586de1ee5eb693e62c95
where the actual Tx start was delayed by TxChainDelay.

Also set already the RADIO STATE register to TX
as soon as the TX START is triggered instead of postponing it
by that microsecond.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>