Convert to an extension to the native simulator
This commit overall refactors the HW models to
be an extension to the native simulator
and as such dependent on it.
Those who desire to continue using them standalone
should weight the effort of switching to this branch
vs staying in the old one,
and the effort of staying on their old framework
vs switching their platform to the native simulator.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>