RADIO: Handle MAXLEN and other related improvements
Cap the Tx length by MAXLEN when doing a Tx.
(Note that just like in real HW, the transmitted length
field is not altered, so the actual Tx'ed packet is corrupted)

During Rx, handle properly MAXLEN:
Instead of erroring out when the received length is more than
MAXLEN, properly truncate the length during reception.

Check PCNF1 configuration for BLE modulation, to ensure
we only got a configuration we support.
(As other configurations are ignored).

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>