standalone nrfx_config: Provide needed definitions for nrf53
When built standalone, the HAL for the nrf53 needs quite
a lot of definitions. Instead of providing our own,
let's just use the template versions provided by the HAL.
The integrator should ensure they are ok, or provide their own.

Note that for Zephyr builds, this file is not used.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>