irq_ctrl: Fix interrupt name getter
When this function was changed to support multiple CPUs,
the check of the interrupt being number being valid was not
updated, which resulted in it thinking too often
the interrupt was out of range, and returning
NULL as the interrupt name string.

Fix it.

This bug only affected debug logs, typically with verbosity
set relatively high.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>