UART: BugFix: Support STARTTx even if Tx is not Off
The real HW seems to support the STARTTX task being triggered
while the Tx is not off, while a DMA copy is not ongoing.
Either while it is just Idle, in the middle of a frame,
or while stopping.
The model assumed this was not supported by the HW and prevented
it.
So let's support it also in the model.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>