UART: Bugfix: Support STARTRX while Rx is stopping
It is also a valid usecase for a STARTRX to be generated
while the UART is still stopping.
In that case it seems the UART just aborts the Rx TO
and handles the STARTRX normally otherwise.

Also related, the HW seems to not generate a RXTO EVENT
if at the Rx TO end a shortcut from ENDRX triggeres a
TASK_STARTRX.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>