nhw_convert_periph_base_addr: Fix include for nrf5340
For the nrf5340 the nrf.h does not provide all peripheral
types but only for the core being built.
We need to use the NHW_peri_types.h header instead
which provides all peripheral types for the whole SOC we
are building.
Similarly nrf_bsim_redef.h header which is meant to hack the
nrf definitions and only provides declarations for the
registers structures for a given core. So we need to declare
the ones that are exclusive for either core here.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>