RADIO: Add test interface to offset Tx and Rx power
Add 2 test specific functions which can be used
to apply a gain to the Tx and Rx power.
The Tx power will be increased by that amount of dBs.
And so will the reported RSSI value.

Note that for Rx is only the RSSI measurement that is changed.
Performance will still be based on the power on the "antenna".

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>