AES_CCM: Avoid UBSAN pointer aligmnet warning when reading CNFPTR data
The area pointed by CNFPTR does not need to be 64bit aligned, even
thought we were reading the packet counter value as if it were.

This is in general not a problem for x86 hosts as non aligned accesses
work fine, but UBSAN warns about it. So let's avoid doing so.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>